requirements.txt
django
djangorestframework
psycopg2-binary
pillow
dj-config-url
drf-yasg
django-filter